Move nested routes further down.
